Restructure source directory
