+++ src/core/log.cpp
