+++ src/core/log.h
