+++ src/core/service.h
@@ -17,41 +17,40 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#ifndef _SESSION_H_
-#define _SESSION_H_
+#ifndef _SERVICE_H_
+#define _SERVICE_H_
 
-#include <ctime>
-#include <memory>
+#include <list>
 #include <boost/asio/io_context.hpp>
+#include <boost/asio/ssl.hpp>
 #include <boost/asio/ip/udp.hpp>
-#include <boost/asio/steady_timer.hpp>
-#include "config.h"
+#include "authenticator.h"
+#include "session/udpforwardsession.h"
 
-class Session : public std::enable_shared_from_this<Session> {
-protected:
+class Service {
+private:
     enum {
-        MAX_LENGTH = 8192,
-        SSL_SHUTDOWN_TIMEOUT = 30
+        MAX_LENGTH = 8192
     };
     const Config &config;
-    uint8_t in_read_buf[MAX_LENGTH];
-    uint8_t out_read_buf[MAX_LENGTH];
-    uint8_t udp_read_buf[MAX_LENGTH];
-    uint64_t recv_len;
-    uint64_t sent_len;
-    time_t start_time;
-    std::string out_write_buf;
-    std::string udp_data_buf;
-    boost::asio::ip::tcp::resolver resolver;
-    boost::asio::ip::tcp::endpoint in_endpoint;
+    boost::asio::io_context io_context;
+    boost::asio::ip::tcp::acceptor socket_acceptor;
+    boost::asio::ssl::context ssl_context;
+    Authenticator *auth;
+    std::string plain_http_response;
     boost::asio::ip::udp::socket udp_socket;
+    std::list<std::weak_ptr<UDPForwardSession> > udp_sessions;
+    uint8_t udp_read_buf[MAX_LENGTH];
     boost::asio::ip::udp::endpoint udp_recv_endpoint;
-    boost::asio::steady_timer ssl_shutdown_timer;
+    void async_accept();
+    void udp_async_read();
 public:
-    Session(const Config &config, boost::asio::io_context &io_context);
-    virtual boost::asio::ip::tcp::socket& accept_socket() = 0;
-    virtual void start() = 0;
-    virtual ~Session();
+    Service(Config &config, bool test = false);
+    void run();
+    void stop();
+    boost::asio::io_context &service();
+    void reload_cert();
+    ~Service();
 };
 
-#endif // _SESSION_H_
+#endif // _SERVICE_H_
