+++ src/core/version.cpp
