+++ src/core/version.h
