+++ src/session/clientsession.cpp
@@ -18,9 +18,9 @@
  */
 
 #include "clientsession.h"
-#include "trojanrequest.h"
-#include "udppacket.h"
-#include "sslsession.h"
+#include "proto/trojanrequest.h"
+#include "proto/udppacket.h"
+#include "ssl/sslsession.h"
 using namespace std;
 using namespace boost::asio::ip;
 using namespace boost::asio::ssl;
