+++ src/session/clientsession.h
