+++ src/session/forwardsession.h
