+++ src/session/serversession.cpp
@@ -18,8 +18,8 @@
  */
 
 #include "serversession.h"
-#include "trojanrequest.h"
-#include "udppacket.h"
+#include "proto/trojanrequest.h"
+#include "proto/udppacket.h"
 using namespace std;
 using namespace boost::asio::ip;
 using namespace boost::asio::ssl;
